requirements.txt
astroid==2.8.2
click==8.1.3
distlib==0.3.4
filelock==3.7.1
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
importlib-metadata==4.12.0
isort==5.9.3
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.6.0
MarkupSafe==2.1.1
mccabe==0.6.1
numpy==1.21.2
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
platformdirs==2.4.0
PyJWT==2.6.0
pylint==2.11.1
six==1.15.0
SQLAlchemy==1.4.41
toml==0.10.2
typing-extensions==3.10.0.2
virtualenv==20.15.1
Werkzeug==2.2.2
wrapt==1.12.1
zipp==3.8.1
